requirements.txt
absl-py==0.9.0
astor==0.8.1
astunparse==1.6.3
atari-py==0.1.15
atomicwrites==1.3.0
attrs==19.1.0
audioread==2.1.8
blinker==1.4
brotlipy==0.7.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
colorama==0.4.1
cryptography==2.9.2
cycler==0.10.0
debtcollector==2.2.0
decorator==4.4.2
future==0.18.2
gast==0.2.2
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
gym==0.17.2
h5py==2.10.0
idna==2.10
image-utils==0.1.6
imageio==2.9.0
importlib-metadata==1.7.0
importlib-resources==3.0.0
iso8601==0.1.12
joblib==0.16.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
librosa==0.6.3
lime==0.2.0.1
llvmlite==0.29.0
Markdown==3.2.2
matplotlib==3.3.1
more-itertools==7.0.0
netaddr==0.8.0
netifaces==0.10.9
networkx==2.4
numba==0.44.0
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.40
opt-einsum==3.3.0
oslo.i18n==5.0.0
oslo.utils==4.5.0
packaging==20.4
pandas==1.1.0
pbr==5.5.0
Pillow==7.2.0
pluggy==0.12.0
progressbar==2.5
protobuf==3.13.0
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
pyglet==1.5.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==4.6.1
python-cephlibs==0.94.5.post1
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
resampy==0.2.1
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.10.1
six==1.15.0
sklearn==0.0
tensorboard==1.15.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.5
termcolor==1.1.0
tf-explain==0.2.1
threadpoolctl==2.1.0
tifffile==2020.8.13
tqdm==4.48.2
urllib3==1.25.10
wcwidth==0.1.7
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
